refactor(core): restructured the `errors` package to make room for new features
