+++ packages/core/src/errors/ErrorSerialiser.ts
@@ -1,29 +1,12 @@
 import { ensure, isDefined, isFunction, JSONObject, JSONValue } from 'tiny-types';
-
-export interface SerialisedError extends JSONObject {
-    /**
-     *  Name of the constructor function used to instantiate
-     *  the original {@apilink Error} object.
-     */
-    name:    string;
-
-    /**
-     *  Message of the original {@apilink Error} object
-     */
-    message: string;
-
-    /**
-     *  Stack trace of the original {@apilink Error} object
-     */
-    stack:   string;
-}
+import { AssertionError, ConfigurationError, ImplementationPendingError, LogicError, TestCompromisedError, TimeoutExpiredError, UnknownError } from './model';
 
 /**
  * @group Errors
  */
 export class ErrorSerialiser {
     private static readonly recognisedErrors: Array<new (...args: any[]) => Error> = [
-        // Built-in JavaScript Errors
+        // Built-in JavaScript errors
         Error,
         EvalError,
         RangeError,
@@ -31,6 +14,15 @@ export class ErrorSerialiser {
         SyntaxError,
         TypeError,
         URIError,
+
+        // Serenity/JS runtime errors
+        AssertionError,
+        ConfigurationError,
+        ImplementationPendingError,
+        LogicError,
+        TestCompromisedError,
+        TimeoutExpiredError,
+        UnknownError,
     ];
 
     static serialise(error: Error): string {
@@ -103,4 +95,22 @@ export class ErrorSerialiser {
 
         return ErrorSerialiser.deserialise({ name, message: message.trim(), stack });
     }
+}
+
+interface SerialisedError extends JSONObject {
+    /**
+     *  Name of the constructor function used to instantiate
+     *  the original {@apilink Error} object.
+     */
+    name:    string;
+
+    /**
+     *  Message of the original {@apilink Error} object
+     */
+    message: string;
+
+    /**
+     *  Stack trace of the original {@apilink Error} object
+     */
+    stack:   string;
 }
