+++ packages/core/src/errors/model/AssertionError.ts
@@ -1,0 +1,31 @@
+import { JSONObject } from 'tiny-types';
+
+import { ErrorSerialiser } from '../ErrorSerialiser';
+import { RuntimeError } from './RuntimeError';
+
+/**
+ * Thrown to indicate that an assertion has failed.
+ *
+ * @group Errors
+ */
+export class AssertionError extends RuntimeError {
+
+    static fromJSON(serialised: JSONObject): AssertionError {
+        const error = new AssertionError(
+            serialised.message as string,
+            ErrorSerialiser.deserialise(serialised.cause as string | undefined),
+        );
+
+        error.stack = serialised.stack as string;
+
+        return error;
+    }
+
+    /**
+     * @param message - Human-readable description of the error and the difference between the expected and actual values
+     * @param [cause] - The root cause of this {@apilink RuntimeError}, if any
+     */
+    constructor(message: string, cause?: Error) {
+        super(AssertionError, message, cause);
+    }
+}
