+++ packages/core/src/errors/model/LogicError.ts
@@ -1,17 +1,21 @@
 import { JSONObject } from 'tiny-types';
 
-import { ErrorSerialiser } from './ErrorSerialiser';
+import { ErrorSerialiser } from '../ErrorSerialiser';
 import { RuntimeError } from './RuntimeError';
 
 /**
- * Thrown to indicate that a test framework or test suite configuration error occurred.
+ * Thrown to indicate that an {@apilink Interaction}, a {@apilink Task} or a test scenario
+ * can't be executed due to a logical error.
+ *
+ * For example, it's not possible to assert on the last HTTP Response if the request
+ * hasn't been performed yet.
  *
  * @group Errors
  */
-export class ConfigurationError extends RuntimeError {
+export class LogicError extends RuntimeError {
 
-    static fromJSON(serialised: JSONObject): ConfigurationError {
-        const error = new ConfigurationError(
+    static fromJSON(serialised: JSONObject): LogicError {
+        const error = new LogicError(
             serialised.message as string,
             ErrorSerialiser.deserialise(serialised.cause as string | undefined),
         );
@@ -26,6 +30,6 @@ export class ConfigurationError extends RuntimeError {
      * @param [cause] - The root cause of this {@apilink RuntimeError}, if any
      */
     constructor(message: string, cause?: Error) {
-        super(ConfigurationError, message, cause);
+        super(LogicError, message, cause);
     }
 }
